feat(op): add unit tests for capability, kernel, storage, and execution tag functionalities
